Create README.md
Make it a habit to regularly back up your wallet onto an external hard drive, USB, or even a piece of paper and store it securely in a separate physical location. By doing so, you’ll have a fail-safe option to restore your lost Bitcoin and avoid the heartache that comes with irreversible loss. But In the case of loss, please ensure to contact Wizard Ghost Champion  for help.email ghostchampionwizard@gmail.com